feat: Refactor ItemList component and add Item type
The ItemList component has been refactored to improve its functionality. Additionally, the Item type has been added to provide better type safety and code readability.
