+++ src/components/ItemList/index.ts
@@ -1,1 +1,3 @@
 export { default as ItemList } from './ItemList';
+export type { ItemProps } from './ItemList';
+export { Item } from './ItemList';
