+++ src/pages/HomePage.tsx
@@ -1,34 +1,7 @@
 import '@styles/index.scss';
 import '../App.css';
 
-import { useNavigate } from 'react-router-dom';
-import { Icon, ItemList, PageHeader, Tab } from '@components';
-
-type HomeItemProps = {
-  title: string;
-  description: string;
-  icon?: string;
-  route?: string;
-};
-
-function HomeItem(props: HomeItemProps) {
-  const navigate = useNavigate();
-
-  return (
-    <div
-      className="box-shadow-medium d-flex p-s-200"
-      onClick={() => props.route && navigate(props.route)}
-    >
-      <div className="col-5 p-s-100">
-        {props.icon && <Icon className="mr-s-100">{props.icon}</Icon>}
-      </div>
-      <div className="col-7">
-        <h4 className="mb-s-100">{props.title}</h4>
-        <p>{props.description}</p>
-      </div>
-    </div>
-  );
-}
+import { Item, ItemList, PageHeader, Tab } from '@components';
 
 function HomePage() {
   const tabs = [
@@ -38,7 +11,7 @@ function HomePage() {
         <ItemList
           items={[
             <>
-              <HomeItem
+              <Item
                 key="offer"
                 title={'Ofertar carona'}
                 description={
@@ -49,7 +22,7 @@ function HomePage() {
               />
             </>,
             <>
-              <HomeItem
+              <Item
                 key="volunteer"
                 title={'Candidatar-se como voluntário'}
                 description={
